redesigned interface to all fit on one screen
